Feature #4409: Mac OSX now reports peak RSS. Also changed type to allow larger memory sizes.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@4682 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Base/askap/current/askap/StatReporter.h
@@ -65,11 +65,11 @@ namespace askap {
             // @return The integer (i.e. the first token) or -1 in the case the
             //          first token was not an integer or the second token was
             //          not the string "kB".
-            int parseValue(std::ifstream& file);
+            long parseValue(std::ifstream& file);
 
             // Convert "val" to a string
             // @return The value as a string with " MB" appended to the end.
-            static std::string kbToMb(int val);
+            static std::string kbToMb(long val);
 
             // Casa timer to measure process execution times
             casa::Timer itsTimer;
